Updated travis integration
