+++ n2v/csf_graph/__init__.py
@@ -1,0 +1,3 @@
+from .csf_graph import CSFGraph
+
+__all__ = ["CSFGraph"]
