+++ tests/__init__.py
@@ -1,1 +1,2 @@
 
+import silence_tensorflow
